pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "intelfactor-inference"
version = "0.1.0"
description = "Edge-native manufacturing inspection + closed-loop RCA engine"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "Tony Lo", email = "tony@intelfactor.ai"},
    {name = "Shreyas Bhardwaj"},
]

# Core dependencies (runs everywhere)
dependencies = [
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# Jetson station node (llama.cpp + TensorRT)
jetson = [
    "llama-cpp-python>=0.2.50",
    "opencv-python-headless>=4.8",
    "flask>=3.0",
]
# GPU server / site hub (vLLM + Triton)
server = [
    "vllm>=0.4",
    "opencv-python-headless>=4.8",
    "flask>=3.0",
    "psycopg2-binary>=2.9",
    "minio>=7.2",
]
# Development + testing
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.3",
]

[project.scripts]
intelfactor-station = "packages.inference.cli:run_station"
intelfactor-hub = "packages.inference.cli:run_hub"

[tool.setuptools.packages.find]
include = ["packages*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 120
